insurance_claim_processor/requirements.txt
google-generativeai
Pillow
Flask
Werkzeug
flask-cors